Time Format Issues fixed

+++ app/src/main/java/com/example/weatherlicious/data/model/forecastweather/ForecastDay.kt
@@ -2,7 +2,7 @@ package com.example.weatherlicious.data.model.forecastweather
 
 data class ForecastDay(
     val date: String,
-    val date_epoch: Int,
+    val date_epoch: Long,
     val day: Day,
     val astro: Astro,
     val hour: List<Hour>
